Invert test class names
